+++ src/main/java/com/starforceps/starforceps/domain/user/api/LoginController.java
@@ -7,34 +7,31 @@ import com.starforceps.starforceps.domain.user.dto.KakaoTokenResponseDto;
 import com.starforceps.starforceps.domain.user.dto.LogoutResponseDto;
 import com.starforceps.starforceps.domain.user.dto.SimpleUserInfoDto;
 import com.starforceps.starforceps.domain.user.dto.UserInfo;
-import com.starforceps.starforceps.domain.user.utils.UserUtils;
+import com.starforceps.starforceps.domain.user.utils.UserUtilsWithKakao;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.*;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.RestTemplate;
 
 @RestController
 @RequiredArgsConstructor
 public class LoginController {
     private final UserService userService;
-    private final UserUtils userUtils;
+    private final UserUtilsWithKakao userUtilsWithKakao;
 
     @GetMapping("/oauth")
     public ResponseEntity<ResponseDto<TokenBundle>> loginCallback(@RequestParam String code) {
-        ResponseEntity<KakaoTokenResponseDto> kakaoTokenResponse = userUtils.getKakaoTokenResponse(code);
+        ResponseEntity<KakaoTokenResponseDto> kakaoTokenResponse = userUtilsWithKakao.getKakaoTokenResponse(code);
         String accessToken = kakaoTokenResponse.getBody().getAccess_token();
         String refreshToken = kakaoTokenResponse.getBody().getRefresh_token();
 
         //유저 정보 조회
-        SimpleUserInfoDto simpleUserInfoDto = userUtils.getUserInfo(accessToken);
+        SimpleUserInfoDto simpleUserInfoDto = userUtilsWithKakao.getUserInfo(accessToken);
         UserInfo userInfo = UserInfo.from(simpleUserInfoDto);
         Long socialId = userInfo.getSocialId();
 
@@ -54,7 +51,7 @@ public class LoginController {
 
     @PostMapping("/oauth/logout")
     public ResponseEntity<ResponseDto<LogoutResponseDto>> logout(HttpServletRequest request) {
-        LogoutResponseDto logoutBody = userUtils.getLogoutResponseDto(request);
+        LogoutResponseDto logoutBody = userUtilsWithKakao.getLogoutResponseDto(request);
         ResponseDto<LogoutResponseDto> response = new ResponseDto<>(
                 2001,
                 "로그아웃 성공",
